fix: crash when calling `OpenProcess()`
Turns out commit f7f0549 did not fully fix the crash. It did add the correct access rights to the `OpenProcess()` call, but PROCESS_TERMINATE requires additional privileges not available by default, even when running xterminate as an administrator. This commit implements the `AdjustTokenPrivileges` API to gain the required privileges for PROCESS_TERMINATE, which should fix the crash permanently.