fix(nsis): error when uninstalling via settings app
Fixes an error displayed by the Windows settings app when attempting to uninstall xterminate.

Due to differences in how the control panel and settings app interpret whitespace in the UninstallString registry key - the settings app treats whitespace as delimiters, while the control panel includes them as part of the path - the error triggers only when using the settings app to uninstall xterminate.

By enclosing the path in double-quotes, both the control panel and the settings app recognize the whitespace as part of the path.